feat(version-2.7): change STEP_INTERVAL to double

+++ src/main/java/jsclub/codefest/sdk/Constants.java
@@ -1,5 +1,5 @@
 package jsclub.codefest.sdk;
 
 public class Constants {
-    public static final int STEP_INTERVAL = 0.5;
+    public static final double STEP_INTERVAL = 0.5;
 }
